+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.0.0] - 2026-01-10
 
 ### Added
-- Initial release of Ada1Status integration
+- Initial release of Adap1Status integration
 - Config flow for easy setup via Home Assistant UI
 - Support for 5 sensor types:
   - Voltage (V)